requirements.txt
streamlit
streamlit-drawable-canvas==0.8.0
matplotlib
Pillow
numpy
